requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
docutils==0.17.1
Flask==2.0.1
idna==3.2
imagesize==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pydocstyle==6.1.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
requests==2.26.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.7
Werkzeug==2.0.1
